@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+		<TargetFramework>net7.0</TargetFramework>
 		<Authors>Shuang Yu</Authors>
 		<Company>AhDai</Company>
 		<Version>1.0.0</Version>
@@ -19,6 +19,7 @@
 		<RepositoryUrl>https://github.com/doublefish/AhDai</RepositoryUrl>
 		<RepositoryType>git</RepositoryType>
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+		<OutputType>Library</OutputType>
 	</PropertyGroup>
 
 	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -30,39 +31,26 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="AhDai.Base" Version="1.0.0" />
+		<PackageReference Include="AhDai.Base" Version="1.0.1" />
 		<PackageReference Include="AhDai.DbContext.MySql" Version="1.0.0" />
 		<PackageReference Include="AhDai.DbContext.Oracle" Version="1.0.0" />
 		<PackageReference Include="AhDai.DbContext.SqlClient" Version="1.0.0" />
 		<PackageReference Include="AhDai.DbContext.SQLite" Version="1.0.0" />
-		<PackageReference Include="log4net" Version="2.0.13" />
-		<PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Abstractions" Version="2.2.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.0.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.0.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Routing" Version="2.2.2" />
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Localization" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Primitives" Version="6.0.0" />
+		<PackageReference Include="log4net" Version="2.0.15" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.14" />
+		<PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.14" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.14" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.14" />
 		<PackageReference Include="Microsoft.Win32.Registry" Version="5.0.0" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-		<PackageReference Include="RabbitMQ.Client" Version="6.2.2" />
-		<PackageReference Include="StackExchange.Redis" Version="2.2.88" />
-		<PackageReference Include="System.Drawing.Common" Version="6.0.0" />
-		<PackageReference Include="System.Management" Version="6.0.0" />
-		<PackageReference Include="System.Net.NetworkInformation" Version="4.3.0" />
-		<PackageReference Include="System.Text.Json" Version="6.0.0" />
-		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
+		<PackageReference Include="RabbitMQ.Client" Version="6.8.0" />
+		<PackageReference Include="StackExchange.Redis" Version="2.7.4" />
+		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+		<PackageReference Include="System.Drawing.Common" Version="7.0.0" />
+		<PackageReference Include="System.Management" Version="7.0.2" />
 	</ItemGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)' == 'net5.0' ">
-		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.14" />
-	</ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)' == 'net6.0' ">
-		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="6.0.2" />
+	<ItemGroup Condition="'$(TargetFramework)' == 'net7.0' ">
 	</ItemGroup>
 
 	<ItemGroup>
