@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Worker">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <Nullable>disable</Nullable>
     <ImplicitUsings>disable</ImplicitUsings>
     <UserSecretsId>dotnet-AhDai.Test-79cf2b10-4435-415e-a28b-65f8c24a0d29</UserSecretsId>
@@ -11,9 +11,5 @@
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
     <PackageReference Include="Elasticsearch.Net" Version="7.17.4" />
     <PackageReference Include="NEST" Version="7.17.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\AhDai.Core\AhDai.Core.csproj" />
   </ItemGroup>
 </Project>
