@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace AhDai.Core.Configs
+namespace AhDai.Core.Configs;
+
+/// <summary>
+/// LogRequestConfig
+/// </summary>
+public class LogRequestConfig
 {
 	/// <summary>
-	/// LogRequestConfig
+	/// 方法
 	/// </summary>
-	public class LogRequestConfig
-	{
-		/// <summary>
-		/// 方法
-		/// </summary>
-		public string[] Methods { get; set; }
-	}
+	public string[] Methods { get; set; }
 }
